MSI: 387: Adapt `\Magento\CatalogInventory\Model\ResourceModel\Stock\Status::addStockDataToCollection`
-- refactoring
